Improved decoupling for Image class
Decoupled ImageLoader classes from Image class; in case there is a new image loader it can be added via the ImageLoader interface (similar to the current one, ImageLoaderSTBI).
Warning: It appears there is a minor memory leak in one of the classes (I think it might be the std::string variable inside the ImageLoaderSTBI class).