Improved Debug class; added #define for glew
Modified printMessage and throwException so that they can take any class instance via templates. This might be a problem later on if it is used everywhere (at least two functions get generated as there is also a getClassName function that gets the class name of the instance).

Added a preprocessor define for glew (GLEW_STATIC). Without this definition, the program halt and logs linker errors.