Added functionality for moving around
UserControls contains code for moving around the world. So far, only the WASD movement is implemented, still need to implement rotation based on mouse movement. CommandHistory is used for storing inputs (key/button presses). FPSManager is used for calculating the FPS of the application (used on frame start and end in main() function). Allows any component to get current application FPS. KeyPressCallbacks was removed due to it being replaced by the UserControls class.