Updated UserControls; enabled flying around
Implemented rotation for UserControls. User is now able to go around the world using the WASD keys and also rotating the camera by pressing the LMB on their mouse and draging the mouse. Currently does not work properly (rotation is not calculated properly), but it can still provide useful information.