Added FPSManager; preparation for next step
Added FPSManager, it allows control over FPS (not fully functional, but works for the most part). FrameTime is impacting user camera movement when using the FPSManager (its alright for <= 500 FPS, above it slows down a lot).

Added player camera to tile manager. This is temporary, there is going to be a seperate User class that keeps track of the players position. The TileManager class is going to be connected to the User class and will generate new terrain depending on the users position in the world (as they get closer to the edge of the world, new world starts to get generated).