Restructured TileManager; slight performance and clarity improvements
Transferred camera user information from TileManager to a separate class (Controller class). The TileManager is subscribed to the Controller class and will get notified when the user moves the camera.

Renamed class used for registering user inputs (class UserControls) to class UserInputRegistry.

Slight memory optimizations to the Tile class (should be impactful for bigger scenes where there are tons of instances of the Tile class).

Moved the projection matrix calculation to the Camera class. This further improves clarity in the main() function.

TO-DO: Transfer user controls to the Controller class; let the UserInputRegistry only work on registering/saving pressed inputs.